Addresses PR comments
